requirements.txt
# Core GUI Framework
PyQt5==5.15.9
PyQt5-sip==12.13.0

# GitHub Integration
PyGithub==2.1.1
GitPython==3.1.40

# HuggingFace Integration
huggingface-hub==0.20.0
transformers==4.35.0
torch==2.1.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
cryptography==41.0.0
Pygments==2.17.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.0
